First commit
Development of a module to obtain the data through UART serial port. In this phase with the PTTY console (it is the one that works correctly; with moserial and minicom I have not managed to make it work) it is possible to enter the value of the keyboard in ASCII. You have to think that the data is only for numbers and in binary.