wrong sitm
the sitm was completely wrong. The condition to send the message is formed by busy (you can send message when it is not busy sending a load) and by a final state of the fsm (idle does not work, but Byte_16 does). Well, it was 5'd6, who knows what it was, and of course, it split the messages randomly.